refactor: imporve some layout of control widgets
