+++ src/containers/DispatchPage/ControlWidget/ControlWidget.module.scss
@@ -1,7 +1,6 @@
 .roomInfoBar {
     position: relative;
     top: 5px;
-    margin-right: 30px;
     display: inline-block;
     height: 32px;
 }
