+++ src/containers/ExamPage/ControlWidget/ControlWidget.module.scss
@@ -3,7 +3,7 @@
   // top:0px;
   z-index: 99;
   display: flex;
-  padding: 10px 50px 10px 20px;
+  padding: 10px 10px 10px 20px;
   background-color:#001529;
   width: 100%;
   height: 50px;
